feat: 할일 갯수 세기 기능 추가
